+++ tests/unit/hbar_test.py
@@ -21,6 +21,24 @@ def test_constructor():
     assert hbar3.to_tinybars() == 50_000_000
     assert hbar3.to_hbars() == 0.5
 
+@pytest.mark.parametrize(
+    'invalid_amount',
+    ['1', True, False, {}, object] 
+)
+def test_constructor_invalid_amount_type(invalid_amount):
+    """Test creation with invalid type raise errors."""
+    with pytest.raises(TypeError, match="Amount must be of type int, float, or Decimal"):
+        Hbar(invalid_amount)
+
+@pytest.mark.parametrize(
+    'invalid_amount',
+    [float('inf'), float('nan')] 
+)
+def test_constructor_non_finite_amount_value(invalid_amount):
+    """Test creation raise errors for non finite amount."""
+    with pytest.raises(ValueError, match="Hbar amount must be finite"):
+        Hbar(invalid_amount)
+
 def test_constructor_with_tinybar_unit():
     """Test creation with unit set to HbarUnit.TINYBAR."""
     hbar1 = Hbar(50, unit=HbarUnit.TINYBAR)
@@ -271,3 +289,12 @@ def test_hash_consistency_for_equal_values():
     assert d[h1] == "value2"
 
     
+
+@pytest.mark.parametrize(
+    'invalid_tinybars',
+    ['1', 0.1, Decimal('0.1'), True, False, object, {}]
+)
+def test_from_tinybars_invalid_type_param(invalid_tinybars):
+    """Test from_tinybar method raises error if the type is not int."""
+    with pytest.raises(TypeError, match=re.escape("tinybars must be an int.")):
+        Hbar.from_tinybars(invalid_tinybars)